refactor[#199] : CancellationCheckModal 파일 오타 수정
